Restructure the project. Add new security module
