+++ GiftCertificatesSystem/security/src/main/java/com/epam/esm/config/MethodSecurityConfig.java
