+++ GiftCertificatesSystem/security/src/main/java/com/epam/esm/exception/ValidateJwtTokenException.java
