+++ GiftCertificatesSystem/security/src/main/java/com/epam/esm/service/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package com.epam.esm.service.impl;
+package com.epam.esm.service;
 
 import lombok.Builder;
 import lombok.Getter;
