+++ GiftCertificatesSystem/security/src/main/java/com/epam/esm/util/CustomPermissionEvaluator.java
