+++ GiftCertificatesSystem/security/src/main/java/com/epam/esm/util/TernaryPredicate.java
