--- GiftCertificatesSystem/service/src/main/java/com/epam/esm/config/ServiceConfig.java
@@ -1,9 +1,0 @@
-package com.epam.esm.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan("com.epam.esm.service")
-public class ServiceConfig {
-}
