+++ GiftCertificatesSystem/settings.gradle
@@ -3,4 +3,5 @@ rootProject.name = 'GiftCertificatesSystem'
 include 'persistence'
 include 'web'
 include 'service'
+include 'security'
